requirements.txt
Django>=3.2,<4.0
psycopg2-binary
redis
